移除user的count接口； 添加account的count接口
