+++ hypervisor/src/main/java/io/leafage/basic/hypervisor/controller/AccountController.java
@@ -65,6 +65,23 @@ public class AccountController {
     }
 
     /**
+     * 统计记录数
+     *
+     * @return 查询到的数据，异常时返回204状态码
+     */
+    @GetMapping("/count")
+    public ResponseEntity<Mono<Long>> count() {
+        Mono<Long> count;
+        try {
+            count = accountService.count();
+        } catch (Exception e) {
+            logger.error("Count user occurred an error: ", e);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.ok(count);
+    }
+
+    /**
      * 根据 username 查询
      *
      * @param username 账号
