+++ hypervisor/src/main/java/io/leafage/basic/hypervisor/service/impl/UserServiceImpl.java
@@ -38,11 +38,6 @@ public class UserServiceImpl extends AbstractBasicService implements UserService
     }
 
     @Override
-    public Mono<Long> count() {
-        return userRepository.count();
-    }
-
-    @Override
     public Mono<UserVO> create(UserDTO userDTO) {
         User user = new User();
         BeanUtils.copyProperties(userDTO, user);
